fix: adjusted controller-mocks service and types and moved mocks
