--- backend/src/controller-mocks/template.mock.ts
@@ -1,7 +1,0 @@
-// import { ControllerEndpointMocks } from './types';
-
-// export const exampleMocks = {
-//   '/route': {
-//     get: {},
-//   },
-// } satisfies ControllerEndpointMocks;
