+++ backend/src/services/controller.service.ts
@@ -4,6 +4,7 @@ import { PrismaClient } from '@prisma/client';
 import { MockService } from '@/controller-mocks/services/mock.service';
 import { MockOptions } from '@/controller-mocks/types';
 import { AxiosRequestConfig } from 'axios';
+import mocks from '@/controllers-mocks';
 
 let apiService: ApiService;
 if (process.env.NODE_ENV === 'production') {
@@ -63,11 +64,15 @@ export interface MockServices extends MockOptions {
 class ControllerService extends MockService<{
   apiService: typeof apiService;
   prisma: typeof prisma;
+  [key: string]: unknown;
 }> {
   public apiService: typeof apiService;
   public prisma: typeof prisma;
 }
 
-const cs = new ControllerService({ prisma, apiService }, { enable: TEST, prefixToRemove: BASE_URL_PREFIX });
+const cs = new ControllerService({ prisma, apiService }, mocks, {
+  enable: TEST,
+  prefixToRemove: BASE_URL_PREFIX,
+});
 
 export default cs;
