+++ backend/src/utils/app-url.ts
@@ -1,0 +1,1 @@
+export const appURL = (path = '') => `${process.env.BASE_URL}${process.env.BASE_URL_PREFIX}${path}`;
